fix: hidden story
