+++ packages/ibm-products/src/components/Tearsheet/_storybook-styles.scss
@@ -31,7 +31,7 @@
 .tearsheet-stories__styles--controls {
   align-items: flex-end;
   display: flex;
-  margin-block-start: 1rem;
+  margin-block: 0.25rem 1rem;
 }
 
 .tearsheet-stories__styles--table {
